+++ src/test/java/kitchenpos/table/domain/OrderTablesTest.java
@@ -9,7 +9,6 @@ import static org.mockito.Mockito.doThrow;
 
 import java.util.Collections;
 import java.util.List;
-import kitchenpos.table.infrastructure.OrderUngroupValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
